unsigned int -> size_t

+++ src/converter/big_endian.h
@@ -62,17 +62,17 @@ inline void verify(const bool cond) {
     }
 }
 
-inline int8_t reads8be(const vector<char> &buf, unsigned int &offs) {
+inline int8_t reads8be(const vector<char> &buf, size_t &offs) {
     verify(buf.size() > offs);
     return buf[offs++];
 }
 
-inline uint8_t readu8be(const vector<char> &buf, unsigned int &offs) {
+inline uint8_t readu8be(const vector<char> &buf, size_t &offs) {
     verify(buf.size() > offs);
     return buf[offs++];
 }
 
-inline be_int16_t reads16be(const vector<char> &buf, unsigned int &offs) {
+inline be_int16_t reads16be(const vector<char> &buf, size_t &offs) {
     verify(buf.size() >= offs + 2);
     int8_t a;
     uint8_t b;
@@ -82,7 +82,7 @@ inline be_int16_t reads16be(const vector<char> &buf, unsigned int &offs) {
     return be_int16_t((a << 8) | b);
 }
 
-inline be_uint16_t readu16be(const vector<char> &buf, unsigned int &offs) {
+inline be_uint16_t readu16be(const vector<char> &buf, size_t &offs) {
     verify(buf.size() >= offs + 2);
     uint8_t a, b;
     a = buf[offs];
@@ -91,7 +91,7 @@ inline be_uint16_t readu16be(const vector<char> &buf, unsigned int &offs) {
     return be_uint16_t((a << 8) | b);
 }
 
-inline be_int32_t reads32be(const vector<char> &buf, unsigned int &offs) {
+inline be_int32_t reads32be(const vector<char> &buf, size_t &offs) {
     verify(buf.size() >= offs + 4);
     int8_t a;
     uint8_t b, c, d;
@@ -103,7 +103,7 @@ inline be_int32_t reads32be(const vector<char> &buf, unsigned int &offs) {
     return be_int32_t((a << 24) | (b << 16) | (c << 8) | d);
 }
 
-inline be_uint32_t readu32be(const vector<char> &buf, unsigned int &offs) {
+inline be_uint32_t readu32be(const vector<char> &buf, size_t &offs) {
     verify(buf.size() >= offs + 4);
     uint8_t a, b, c, d;
     a = buf[offs];
@@ -114,7 +114,7 @@ inline be_uint32_t readu32be(const vector<char> &buf, unsigned int &offs) {
     return be_uint32_t((a << 24) | (b << 16) | (c << 8) | d);
 }
 
-inline vector<char> readbytes(const vector<char> &buf, unsigned int &offs, const unsigned int n) {
+inline vector<char> readbytes(const vector<char> &buf, size_t &offs, const size_t n) {
     assert(n > 0);
     verify(buf.size() >= offs + n);
     vector<char> chars(n);
@@ -123,7 +123,7 @@ inline vector<char> readbytes(const vector<char> &buf, unsigned int &offs, const
     return chars;
 }
 
-inline vector<uint8_t> readu8bytes(const vector<char> &buf, unsigned int &offs, const unsigned int n) {
+inline vector<uint8_t> readu8bytes(const vector<char> &buf, size_t &offs, const size_t n) {
     assert(n > 0);
     verify(buf.size() >= offs + n);
     vector<uint8_t> chars(n);
@@ -132,14 +132,14 @@ inline vector<uint8_t> readu8bytes(const vector<char> &buf, unsigned int &offs, 
     return chars;
 }
 
-inline void copybytes(const vector<char> &buf, char *dst, unsigned int &offs, const unsigned int n) {
+inline void copybytes(const vector<char> &buf, char *dst, size_t &offs, const size_t n) {
     if (n == 0) return;
     verify(buf.size() >= offs + n);
     copy(buf.begin() + offs, buf.begin() + offs + n, dst);
     offs += n;
 }
 
-inline void copyu8bytes(const vector<char> &buf, uint8_t *dst, unsigned int &offs, const unsigned int n) {
+inline void copyu8bytes(const vector<char> &buf, uint8_t *dst, size_t &offs, const size_t n) {
     if (n == 0) return;
     assert(n > 0);
     verify(buf.size() >= offs + n);
